+++ src/foldersearch.ts
@@ -1,6 +1,7 @@
 import * as core from "@actions/core";
 import * as fs from "fs";
 import * as path from "path";
+import { minimatch } from "minimatch";
 
 /**
  * find packages in subfolders that contain a package.json
@@ -17,7 +18,7 @@ async function findPackageJsonFolders(currentPath: string, ignoreFolders: string
         continue;
       }
 
-      if (ignoreFolders.some((folder) => dirent.name.startsWith(folder))) {
+      if (ignoreFolders.some((folder) => minimatch(dirent.name, folder))) {
         core.info(`Skipping folder: ${fullPath} due ignoreFolders setting`);
         continue;
       }
@@ -38,4 +39,35 @@ async function findPackageJsonFolders(currentPath: string, ignoreFolders: string
   return foundFolders;
 }
 
-export { findPackageJsonFolders };
+/**
+ * find folders that contain *.csproj files
+ * @param currentPath the path to start searching
+ * @param ignoreFolders folders to ignore
+ */
+async function findCsProjectFolders(currentPath: string, ignoreFolders: string[]): Promise<string[]> {
+  const dirents = fs.readdirSync(currentPath, { withFileTypes: true });
+  const foundFolders: string[] = [];
+  for (const dirent of dirents) {
+    const fullPath = path.join(currentPath, dirent.name);
+    if (dirent.isDirectory()) {
+      if (fullPath.includes("node_modules") || dirent.name.startsWith(".")) {
+        continue;
+      }
+
+      if (ignoreFolders.some((folder) => minimatch(dirent.name, folder))) {
+        core.info(`Skipping folder: ${fullPath} due ignoreFolders setting`);
+        continue;
+      }
+
+      const files = fs.readdirSync(fullPath);
+      const csprojExists = files.some((file) => file.endsWith(".csproj"));
+
+      if (csprojExists) {
+        foundFolders.push(fullPath);
+      }
+    }
+  }
+  return foundFolders;
+}
+
+export { findPackageJsonFolders, findCsProjectFolders };