+++ src/index.ts
@@ -3,6 +3,7 @@ import * as github from "@actions/github";
 import * as foldersearch from "./foldersearch";
 import * as prcomment from "./prcomments";
 import * as npmlicensecheck from "./npmlicensecheck";
+import * as nugetlicensecheck from "./nugetlicensecheck";
 
 /**
  * The main entry point
@@ -23,6 +24,12 @@ async function run(): Promise<void> {
     // remove old comments
     await prcomment.removeOldPullRequestComments(pullRequestNumber);
 
+    // find all *.csproj folders
+    const csprojFolders = await foldersearch.findCsProjectFolders("./", ignoreFolders);
+
+    // process each folder
+    let textForComment = await nugetlicensecheck.processNuget(csprojFolders);
+
     // find all package.json folders
     const packageJsonFolders = await foldersearch.findPackageJsonFolders("./", ignoreFolders);
 
@@ -30,9 +37,14 @@ async function run(): Promise<void> {
     for (const folder of packageJsonFolders) {
       const currentFolder = process.cwd();
       await process.chdir(folder);
-      await npmlicensecheck.processNpm(folder, pullRequestNumber);
+      textForComment += await npmlicensecheck.processNpm(folder);
       await process.chdir(currentFolder);
     }
+
+    // create comment
+    let prComment = `## License Report\n\n`;
+    prComment += textForComment;
+    await prcomment.writePullRequestComment(prComment, pullRequestNumber);
   } catch (error) {
     if (error instanceof Error) {
       core.setFailed(error);