+++ CHANGELOG.md
@@ -7,7 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### dependabot: \#18 Bump tj-actions/changed-files from 39 to 40
+### Added
+
+- Added support for nuget
+
+### Updated
+
+- Updated readme to use the latest version
+- Bump tj-actions/changed-files from 39 to 40
 
 ## [0.2.1] - 2023-10-26